Fixing cog_state problems, passes tests
One problem was that we were casting the pointer to the list element
rather than the list's data pointer.
Another problem was in the update function, where we were not checking
whether we were in the same state as the transition before advancing.
Both of these problems are now fixed and the test passes.